[feat]: 활동 등록을 위한 2단계 구현시 페이지 분할 하면 첫 단계에 입력한 데이터를 유지할 수 없어서 한 페이지로 합침
